requirements.txt
facebook-business
openai
langchain
langgraph
pinecone
streamlit
pandas
python-dotenv
facebook-business
